udpate Repository

+++ src/main/java/com/zhaoxue/bean_backend/common/bean/repository/PersonRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.zhaoxue.bean_backend.common.bean.entity.PersonEntity;
 
-public interface PersonRepository extends JpaRepository<PersonEntity, Integer> {
+public interface PersonRepository extends JpaRepository<PersonEntity, Long> {
     // 这里可以根据需要定义额外的方法
 }
